ModuleManager no longer owns modules.
Since modules are already singletons, we move ownership of them away
from ModuleManager. This allows all inter-module calls to be statically
dispatched.

Since we no longer need to be able to pick out a specific module
efficiently, we can switch from a `map` to a `vector` for storing the
active list of modules. We also no longer need typeinfo.