Update projectile colors in-flight
When projectiles are deflected, their team is switched and thus their
color should too. The game updates the projectile team for us, so we can
just recalculate the color on-the-fly to pick up the new color.

We don't bother caching the old team name and checking for changes and
so on - there are not that many projectiles on the map at any one time
to actually make it a significant performance bottleneck.

This fixes #84.