Add SrcTV+ detection module
This module is designed for progressive enhancement with
[srctv+](https://github.com/dalegaard/srctvplus) or an equivalent server
plugin(or a future TF2 update).

Normally SourceTV will not send player- and weaponlocal data, as well as
leaving out many relevant game events like `player_hurt` and
`player_healed`. The SrcTV+ server plugin makes it send everything.

For CE to support servers both with and without SrcTV+(both dedicated
and listen servers) we need to be able to detect the presence of the
server plugin. This is what this module does.

Other modules can query the status via the `SrcTVPlus::IsAvailable()`
call, and they can subscribe to status updates using the
`SrcTVPlusListener` class.

Internally detection works by replacing the receive proxy for the
`DT_LocalPlayerExclusive.m_nTickBase` field with a hook. This field,
like the rest of `DT_LocalPlayerExclusive` is normally only sent to the
local player, so we can simply look for updates to it to determine if
SrcTV+ is available. Note however, when the game spawns an entity the
receive proxy is called, sometimes multiple times, for the new entity
and thus we need to look for at least a few updates. We've chosen 10 as
a somewhat arbitrary cutoff point, to balance detection speed with
robustness.

This fixes #87.