Small optimization in `ColorFromString`
When `ColorFromString` performs a scheme color lookup, we always did two
looksup so we could figure out if the color was present or not.

We can actually do a small short-circuit evaluation here, and only do
the second lookup if the color is the color we specified as the default
color in the first lookup. Since this default color is picked as
(254,1,254) it's quite unlikely someone would ever pick it as a real
color.